check code by rubocop

+++ lib/vvm-rb/switcher.rb
@@ -7,14 +7,10 @@ class Switcher
 
   def use
     current = get_current_dir
-    if File.exists?(current)
-      FileUtils.rm(current)
-    end
+    FileUtils.rm(current) if File.exists?(current)
     unless @version == 'system'
       vims_dir = get_vims_dir(@version)
-      unless File.exists?(vims_dir)
-        abort "#{@version} is not installed."
-      end
+      abort "#{@version} is not installed." unless File.exists?(vims_dir)
       FileUtils.ln_s(vims_dir, current)
     end
   end
